Added banner image and fixed duplicate calls to API
